refactor: remove input port directory for auth
